==> R_RMD_Analysis/Step1_Features.Rmd <==
---
title: "Multi-Level Linguistic Alignment in a Dynamic CPS Tasks: Step 1 Features"
date: 03/30/22
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())
```

## What this Notebook Contains

> Notebook takes raw output files from the ALIGN analysis and performs a clean-up to creates necessary variables for analysis, with a focus on creating the composite linguistic variables

#### Step 1: Import main data files and select relevant variables and combine into master datasheet

```{r, warning=FALSE, message=FALSE, error=FALSE}
library(tidyverse)
library(pander)
```

Run the actual data (be sure to update the name of the file to be saved at the end)

```{r}
PennPOS_keep = read.csv("../Data/ALIGN_Output/PennPOS_keep_T2T.txt", sep="\t")
StanPOS_keep = read.csv("../Data/ALIGN_Output/StanPOS_keep_T2T.txt", sep="\t")
PennPOS_remove = read.csv("../Data/ALIGN_Output/PennPOS_remove_T2T.txt", sep="\t")
StanPOS_remove = read.csv("../Data/ALIGN_Output/StanPOS_remove_T2T.txt", sep="\t")
```

Run the baseline data (be sure to update the name of the file to be saved at the end)

```{r}
# PennPOS_keep = read.csv("../Data/ALIGN_Output/PennPOS_keep_T2T_Baseline.txt", sep="\t")
# StanPOS_keep = read.csv("../Data/ALIGN_Output/StanPOS_keep_T2T_Baseline.txt", sep="\t")
# PennPOS_remove = read.csv("../Data/ALIGN_Output/PennPOS_remove_T2T_Baseline.txt", sep="\t")
# StanPOS_remove = read.csv("../Data/ALIGN_Output/StanPOS_remove_T2T_Baseline.txt", sep="\t")
```

Code below runs the same for either actual and baseline

```{r, warning=FALSE, message=FALSE, error=FALSE}
PennPOS_keep = PennPOS_keep %>% select(condition_info, partner_direction, time, utterance_length1, utterance_length2, 
                lexical_lem1, lexical_lem2, lexical_lem3, lexical_tok1, lexical_tok2, lexical_tok3, 
                syntax_penn_lem2, syntax_penn_tok2, syntax_penn_lem3, syntax_penn_tok3, 
                cosine_semanticL)

StanPOS_keep = StanPOS_keep %>% select(condition_info, partner_direction, time, syntax_stan_lem2, syntax_stan_tok2, syntax_stan_lem3, syntax_stan_tok3)

mergekeep = left_join(PennPOS_keep,StanPOS_keep)
```

```{r, warning=FALSE, message=FALSE, error=FALSE}

PennPOS_remove = PennPOS_remove %>% select(condition_info, partner_direction, time, 
                  syntax_penn_lem2, syntax_penn_lem3, syntax_penn_tok2, syntax_penn_tok3) %>% rename(syntax_penn_lem2_nodp = syntax_penn_lem2, syntax_penn_lem3_nodp = syntax_penn_lem3, syntax_penn_tok2_nodp = syntax_penn_tok2, syntax_penn_tok3_nodp = syntax_penn_tok3)

StanPOS_remove = StanPOS_remove %>% select(condition_info, partner_direction, time,
                  syntax_stan_lem2, syntax_stan_lem3, syntax_stan_tok2, syntax_stan_tok3) %>% rename(syntax_stan_lem2_nodp = syntax_stan_lem2, syntax_stan_lem3_nodp = syntax_stan_lem3, syntax_stan_tok2_nodp = syntax_stan_tok2, syntax_stan_tok3_nodp = syntax_stan_tok3)

mergeremove = left_join(PennPOS_remove,StanPOS_remove)
```

```{r, warning=FALSE, message=FALSE, error=FALSE}
df.merge = left_join(mergekeep, mergeremove)
```

#### Step 2: Clean-up for variable creation, relabeling for readability, etc.)

```{r, warning=FALSE, message=FALSE, error=FALSE}
source("../R_Helper/Step1_helper.R")
df = cleanup(df.merge)
```

* Relies heavily on "Step1_helper.R." This code performs many of the more mundane tasks and imports the following files to integrate necessary variables

  * Step1_FilesToIntegrate/SWITCH_PARTICIPANTS.csv
  * Step1_FilesToIntegrate/CPS2_Teams.csv
  * Step1_FilesToIntegrate/Levels_PP_Logs_Lab_Before_Split_by_Trophy.csv

* Important note that CPS2_Teams.csv comes from the official CPS2 folder of processed data and documentation states: "Official list of team IDs" and in here there are only 288 individuals. Six participants (two teams) removed after running Step1_helper.R, problem in that outcomes are not recorded for two teams

```{r}
trophies = read.csv('../Accompanying_Data/Step1_FilesToIntegrate/Levels_PP_Logs_Lab_Before_Split_by_Trophy.csv', sep=",") 
```

```{r}
df = df %>% select(Team, School, Block, Concept, Goal, Level, trophy, level_duration, relative_start_time, revisited, speaker_target, aligner_score, aligner_role, utter_order, utterlen_target, utterlen_aligner, semantic, lexical_lem1, lexical_tok1, syntax_penn_lem2_nodp, syntax_penn_tok2_nodp, syntax_stan_lem2_nodp, syntax_stan_tok2_nodp, syntax_penn_lem3_nodp, syntax_penn_tok3_nodp, syntax_stan_lem3_nodp, syntax_stan_tok3_nodp)
```

#### Step 3: Create composite ling measures as described in manuscript

```{r}
df_composite = df %>% rowwise() %>% 
  mutate(lexical = (lexical_lem1 + lexical_tok1)/2) %>%
  
  mutate(syn2nd = (syntax_penn_lem2_nodp + syntax_penn_tok2_nodp + syntax_stan_lem2_nodp + syntax_stan_tok2_nodp)/4) %>%
  
  mutate(syn3nd = (syntax_penn_lem3_nodp + syntax_penn_tok3_nodp + syntax_stan_lem3_nodp + syntax_stan_tok3_nodp)/4) %>%
  
  mutate(syntax = (syn2nd + syn3nd)/2) %>%
  
  select(-lexical_lem1, -lexical_tok1, -syntax_penn_lem2_nodp, -syntax_penn_tok2_nodp, -syntax_stan_lem2_nodp, -syntax_stan_tok2_nodp, -syntax_penn_lem3_nodp, -syntax_penn_tok3_nodp, -syntax_stan_lem3_nodp, -syntax_stan_tok3_nodp)

pander(head(df_composite[,c(1:6,17:21)]))

pander(names(df_composite))
```

```{r}
write.csv(df_composite, "../Data/Step1_PrepareFeatures.csv", row.names=FALSE)
```

```{r}
# write.csv(df_composite, "../Data/Step1_PrepareFeatures_Baseline.csv", row.names=FALSE)
```
