==> R_RMD_Analysis/Step3_Descriptives.Rmd <==
---
title: "Multi-Level Linguistic Alignment in a Dynamic CPS Tasks: Descriptives"
date: 03/30/22
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())
```

## Notebook contains code for computing level stats on relationship between # of turns, trophy completions, block, and verbosity

#### Step 1: Import main data file and run script to prep variables for analysis

```{r, warning=FALSE, message=FALSE, error=FALSE}
library(tidyverse)
library(pander)
```

```{r}
main.data = read_csv("../Data/Step1_PrepareFeatures.csv", show_col_types = FALSE)
```

Code below just generates a new variable for how many overall turns were present in each level and adds to the main datasheet

```{r, warning=FALSE, message=FALSE, error=FALSE}
turns = main.data %>% group_by(Team, School, Block, Level, revisited) %>%
  summarise(numberturns = n())
main.data = left_join(main.data, turns)
```

# GENERATE WHAT ACTUALLY APPEARS IN PAPER

```{r}
## how many original levels
levels1 = main.data %>% select(Team, School, Block, Level, revisited, trophy) %>% distinct(); nrow(levels1)
```

Number of levels after removing short levels of fewer than 20 turns

```{r}
rmTurns = 20

df.20.more = filter(main.data, numberturns >= rmTurns)
## how many levels after 20 turns, also same data as models
levels2 = filter(main.data, numberturns >= rmTurns) %>% filter(utterlen_aligner < 71 & utterlen_target < 71) %>% select(Team, School, Block, Level, revisited, trophy) %>% distinct(); nrow(levels2)

```

Of the longer levels, number of turns and words per turn

```{r}
# overallturnswords = df.20.more %>% group_by(Team, School, Block, Level, revisited, trophy) %>% 
#   summarize_at(c("numberturns", "utterlen_aligner"), list(mean, sd)); overallturnswords 

overallturnswords = df.20.more %>% filter(utterlen_aligner < 71 & utterlen_target < 71) %>%
  summarize_at(c("numberturns", "utterlen_aligner"), list(mean, sd)); colnames(overallturnswords) <- c("Mturn", "Mutterlen", "SDturn", "SDutterlen"); overallturnswords

```

Of the longer levels, outcomes based on coins

```{r}
## how many trophies distributed across reduced levels
levtab2 = table(levels2$trophy);  
levtab2
```

Generate table showing time and turn information across outcomes, also generate statistical tests comparing these measures for outcomes

```{r}
levBlotab2 = df.20.more %>% filter(utterlen_aligner < 71 & utterlen_target < 71) %>%
  group_by(trophy) %>% 
  summarize_at(c("level_duration", "numberturns"), list(mean = mean, sd = sd)) %>%
  mutate(
    trophy = case_when(
      trophy == "none" ~ "None",
      trophy == "silver" ~ "Silver",
      trophy == "gold" ~ "Gold") %>% 
      as.factor %>% 
      fct_relevel("None", "Silver", "Gold")
  )

```

```{r}
p_value_one <- tibble(
  x = c("None", "None", "Silver", "Silver"),
  y = c(775, 800, 800, 775)
)

p_value_two <- tibble(
  x = c("Silver", "Silver", "Gold", "Gold"),
  y = c(625, 650, 650, 625)
)

fig1 = levBlotab2 %>% 
  ggplot(aes(trophy, level_duration_mean)) +
  geom_col(aes(fill = trophy), color = "black", width = 0.85) +
  geom_errorbar(aes(ymin = level_duration_mean - level_duration_sd,
                    ymax = level_duration_mean + level_duration_sd),
                color = "#22292F",
                width = .1) +
  geom_line(data = p_value_one, 
            aes(x = x, y = y, group = 1)) +
  geom_line(data = p_value_two,
            aes(x = x, y = y, group = 1)) +
  annotate("text", x = 1.5, y = 810, 
           label = "***",
           size = 8, color = "#22292F") +
  annotate("text", x = 2.5, y = 660,
           label = "***",
           size = 8, color = "#22292F") +
  scale_fill_grey(start = 0.3) +
  scale_y_continuous(limits = c(0, 865), expand = c(0, 0)) +
  guides(fill = "none") +
  theme_minimal() +
  labs(
    x = "",
    y = "Mean Level Duration",
    title = "Duration",
    # caption = "Error bars indicate standard deviations"
  ) + 
  
  theme_bw() +
  theme(plot.title=element_text(size=14,face="bold"), # theme_light()
        axis.text.y=element_text(size=12),
        axis.ticks.y=element_blank(),
        axis.text.x=element_text(size=12, face="bold"),
        axis.title=element_text(size=12,face="bold"),
        # plot.caption = element_text(hjust = 0),
        strip.text.y = element_text(hjust=0,vjust = 1,angle=180,face="bold"),
    axis.line = element_line(color = "#3D4852"),
    axis.ticks = element_line(color = "#3D4852"),
    panel.grid.major.y = element_line(color = "#DAE1E7"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.y = element_blank()
  )
```

```{r}

p_value_one <- tibble(
  x = c("Silver", "Silver", "Gold", "Gold"),
  y = c(122, 127, 127, 122)
)

p_value_two <- tibble(
  x = c("None", "None", "Silver", "Silver"),
  y = c(137, 142, 142, 137)
)

fig2 = levBlotab2 %>% 
  ggplot(aes(trophy, numberturns_mean)) +
  geom_col(aes(fill = trophy), color = "black", width = 0.85) +
  geom_errorbar(aes(ymin = numberturns_mean - numberturns_sd,
                    ymax = numberturns_mean + numberturns_sd),
                color = "#22292F",
                width = .1) +
  geom_line(data = p_value_one, 
            aes(x = x, y = y, group = 1)) +
  geom_line(data = p_value_two,
            aes(x = x, y = y, group = 1)) +
  annotate("text", x = 2.5, y = 129, 
           label = "***",
           size = 8, color = "#22292F") +
  annotate("text", x = 1.5, y = 144,
           label = "*",
           size = 8, color = "#22292F") +
  scale_fill_grey(start = 0.3) +
  scale_y_continuous(limits = c(0, 155), expand = c(0, 0)) +
  guides(fill = "none") +
  theme_minimal() +
  labs(
    x = "",
    y = "Mean Turns",
    title = "Turns",
    caption = "*p<.05, ***p<.001"
  ) + 

  theme_bw() +
  theme(plot.title=element_text(size=14,face="bold"), # theme_light()
        axis.text.y=element_text(size=12),
        axis.ticks.y=element_blank(),
        axis.text.x=element_text(size=12, face="bold"),
        axis.title=element_text(size=12,face="bold"),
        plot.caption = element_text(size=12),
        strip.text.y = element_text(hjust=0,vjust = 1,angle=180,face="bold"),
    axis.line = element_line(color = "#3D4852"),
    axis.ticks = element_line(color = "#3D4852"),
    panel.grid.major.y = element_line(color = "#DAE1E7"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.y = element_blank()
  )
fig2
```

```{r}
library(patchwork)
durationPlots = fig1 + fig2
durationPlots
```
```{r}
# ggsave(filename = "Figs/trophy_levels_time.png",width = 8, height = 4.5, dpi = 300, units = "in", device='png')
```

Statistical significance tests


```{r, warning=FALSE, message=FALSE, error=FALSE}
library(lme4)
library(emmeans)
library(multcomp)
library(car)
```

```{r}
levBlotab3 = df.20.more %>% filter(utterlen_aligner < 71 & utterlen_target < 71) %>%
  dplyr::select(1:8, "numberturns") %>% unique()

levBlotab3 = levBlotab3 %>% mutate(TeamRE = paste0(Team, School))
a1 = lmer(level_duration ~ trophy + (1|TeamRE), levBlotab3); summary(a1); Anova(a1, type = 2)
a1.emm <- emmeans(a1, ~ trophy)
pairs(a1.emm, adjust = "none")

## same as above, but cleaner and can be used with lmer if desired
# a1 = aov(level_duration ~ trophy, levBlotab3); summary(a1) 
# pairwise.t.test(levBlotab3$level_duration, levBlotab3$trophy, p.adj = "bonf")

a2 = lmer(numberturns ~ trophy + (1|TeamRE), levBlotab3); summary(a2); Anova(a2, type = 2)
a2.emm <- emmeans(a2, ~ trophy)
pairs(a2.emm, adjust = "none")

```


# DEEPER DIVE INTO DESCRIPTIVE DATA OF PERFORMANCE

#### Step 1

> Get a few stats on relationship between # of turns, trophy completions, and block

```{r}
## how many trophies distributed across all levels; how many blocks?
levtab1 = table(levels1$trophy); blotab1 = table(levels1$Block);

## how many trophies distributed across reduced levels; how many blocks?
levtab2 = table(levels2$trophy);  blotab2 = table(levels2$Block); 

levtab1; levtab2
blotab1; blotab2
```

```{r}
## What levels are dropped, based on TROPHY, when reducing by number of turns?

## how many levels were dropped BY TROPHIES after removing turns with 20 or fewer turns
# levtab1[1]-levtab2[1]; ((levtab1[1]-levtab2[1])/levtab1[1])*100
# levtab1[3]-levtab2[3]; ((levtab1[3]-levtab2[3])/levtab1[3])*100
# levtab1[2]-levtab2[2]; ((levtab1[2]-levtab2[2])/levtab1[2])*100
matA = matrix(c(names(levtab1[1]), levtab1[1], levtab2[1], ((levtab1[1]-levtab2[1])/levtab1[1])*100,
                names(levtab1[3]), levtab1[3], levtab2[3], ((levtab1[3]-levtab2[3])/levtab1[3])*100,
                names(levtab1[2]), levtab1[2], levtab2[2], ((levtab1[2]-levtab2[2])/levtab1[2])*100),
                ncol=4,byrow=TRUE)
colnames(matA) <- c("Trophy", "Original","Retained","% Drop")
matA
```


```{r}
## What levels are dropped, based on BLOCK, when reducing by number of turns?

# blotab1[1]-blotab2[1]; ((blotab1[1]-blotab2[1])/blotab1[1])*100
# blotab1[3]-blotab2[3]; ((blotab1[3]-blotab2[3])/blotab1[3])*100
# blotab1[2]-blotab2[2]; ((blotab1[2]-blotab2[2])/blotab1[2])*100
matB = matrix(c("Warmup", blotab1[3], blotab2[3], ((blotab1[3]-blotab2[3])/blotab1[3])*100,
                "ExpBlock1", blotab1[1], blotab2[1], ((blotab1[1]-blotab2[1])/blotab1[1])*100,
                "ExpBlock2", blotab1[2], blotab2[2], ((blotab1[2]-blotab2[2])/blotab1[2])*100),
                ncol=4,byrow=TRUE)
colnames(matB) <- c("Block", "Original","Retained","% Drop")
matB
```

#### STEP 2

Repeat much of the above (stats on relationship between # of turns, trophy completions, and block) but in a format more easily used in paper and for interpretation

##### how many original levels; broken down by Block and Trophy

```{r}
main.data$trophy = factor(main.data$trophy, levels=c("none", "silver", "gold"))
main.data$Block = factor(main.data$Block, levels=c("Warmup", "ExpBlock1", "ExpBlock2"))

levels1.agg = main.data %>% group_by(Team, School, Block, Level, revisited, trophy) %>%
  summarise_if(is.numeric, list(~ mean(., na.rm = TRUE))); x1 = xtabs(~trophy + Block, levels1.agg)
x1
```

##### how many reduced levels after 20 or fewer turns; broken down by Block and Trophy

```{r}
levels2.agg = filter(main.data, numberturns >= rmTurns) %>% group_by(Team, School, Block, Level, revisited, trophy) %>%
  summarise_if(is.numeric, list(~ mean(., na.rm = TRUE))); x2 = xtabs(~trophy + Block, levels2.agg)
x2
```

##### report above as a percentage (dropped vs. retained)

```{r}
droplevel = ((x1-x2)/x1)*100; droplevel
droplevel2 = ((x2)/x1)*100; droplevel2
```

##### how many levels removed

```{r}
levels2.agg = filter(main.data, numberturns < rmTurns) %>% group_by(Team, School, Block, Level, revisited, trophy) %>%
  summarise_if(is.numeric, list(~ mean(., na.rm = TRUE))); x3 = xtabs(~trophy + Block, levels2.agg)
x3
```

What is the story emerging from above? How does this change any of the analyses attempting to do? As implications when trying to relate language dynamics, that require temporal component of at least 20 turns to develop, to performance outcomes.

```{r}
# First, way more levels are attempted from block to block
colSums(x1)
# But this gets equalized out some after removing levels with 20 or less turns
colSums(x2)
# This shows the reduction. Although more levels are dropped from ExpBlock2, not that dramatic
matB
```

```{r}
# There is also a good distribution of levels with different performance outcomes
rowSums(x1)
# But this gets reduced quite a bit with 20 or less turns
rowSums(x2)
# This shows the reduction. Many more gold levels are dropped, followed by none levels
matA
```

First the raw distribution of levels as trophy by block, overall. Clearly gold trophies really ramp up from Warmup to ExpBlock2, silver trophies stay constant, and none levels increase as well. Much has to do with the fact that more levels are attempted from block to block

```{r}
x1
```

This is a better look as a percentage of overall levels based on those attempted in each block. Interestingly, of those attempted, silver actually decreases while gold goes up. None stays constant.

```{r}
t(t(x1)/colSums(x1))
```

> Now, the raw distribution of levels as trophy by block in the reduced set.

```{r}
x2
```

> And what it looks like in terms of levels as trophy by block for those actually attempted in the reduced set. Sort of same pattern as the original, but just not as pronounced.

```{r}
t(t(x2)/colSums(x2))
```

> Why not as pronounced, because there are a lot of levels dropped in the gold and not so much in silver. The most affected is gold trophies in the later blocks.

> Lots of levels with gold trophies are removed (implication: people are relying on non-verbal cues for these); relatively way fewer silver trophies were removed (in fact, more failed levels than those getting silver: implication: silver levels have lots of language relative to gold levels)

```{r}
droplevel = ((x1-x2)/x1)*100; droplevel
```

<!-- thinking it is not wise to combine gold and silver in terms of relating language to performance, they have their own unique language properties. Clearly, penalizing most the levels where gold trophies were achieved as they had the fewest conversational turns. Very little impact on levels where silver trophies were awarded. -->

> One more look, just confirms that golds really have fewer words. Followed by silver with second most, and none with the most. Also shows trends where the number of words drops off across blocks

```{r, warning=FALSE, message=FALSE, error=FALSE}
levBlotab = main.data %>% group_by(factor(trophy, levels=c("none", "silver", "gold")), factor(Block, levels=c("Warmup", "ExpBlock1", "ExpBlock2"))) %>% summarise(mean(numberturns)); colnames(levBlotab) <- c("Trophy", "Block", "AvgNumTurns"); levBlotab

lev_duration = main.data %>% group_by(factor(trophy, levels=c("none", "silver", "gold")), factor(Block, levels=c("Warmup", "ExpBlock1", "ExpBlock2"))) %>% summarise(mean(level_duration)); colnames(lev_duration) <- c("Trophy", "Block", "AvgDuration"); lev_duration
```

```{r, warning=FALSE, message=FALSE, error=FALSE}
df.20.more = filter(main.data, numberturns >= rmTurns)

levBlotab2 = df.20.more %>% group_by(factor(trophy, levels=c("none", "silver", "gold")), factor(Block, levels=c("Warmup", "ExpBlock1", "ExpBlock2"))) %>% summarise(mean(level_duration), mean(numberturns)); colnames(levBlotab2) <- c("Trophy", "Block", "AvgDuration", "AvgNumTurns"); levBlotab2

```


############# OTHER: CAN BE IGNORED ###############

```{r}
# levels1 = df.sl2 %>% select(Team, School, aligner_score) %>% distinct()
# levels2 = df.sl2 %>% select(Team, School, block_manip, Level, revisited, numberturns) %>% distinct()
# 
# quantile(levels2$numberturns)
# IQR(levels2$numberturns)
# 
# duration = levels2$numberturns
# breaks = seq(5, 200, by=5)
# duration.cut = cut(duration, breaks, right=FALSE)
# duration.freq = table(duration.cut)
# duration.cumfreq = cumsum(duration.freq)
# 
# duration.cumrelfreq = duration.cumfreq / nrow(levels2) *100
# duration.cumrelfreq

```
