==> R_RMD_Analysis/Step3_M1_Lexical.Rmd <==
---
title: "Multi-Level Linguistic Alignment in a Dynamic CPS Tasks: Step 3 Model 1: Lexical"
date: 03/28/22
output: html_document
---

```{r global-options, include=FALSE}
knitr::opts_chunk$set(fig.path='Figs/',
                      warning=FALSE, message=FALSE, error=FALSE)
```
<!-- # echo=FALSE : suppress all the code but show text/results output -->
<!-- # results='hide' : show code but hide the text/results output -->
<!-- # fig.show='hide' : hide figures -->

> Notebook contains code for replicating section in manuscript: Analysis 1: Alignment over time

### Step 1: Import main data file and run script to prep variables for analysis

```{r, warning=FALSE, message=FALSE, error=FALSE}
library(tidyverse)
library(pander)

rm(list=ls())
```

```{r}
###### REAL DATA : BE SURE TO UPDATE NAMES OF THE SAVED OUTPUT FILES ########
main.data = read.csv("../Data/Step1_PrepareFeatures.csv")

###### BASELINE DATA FOR COMPARISON : BE SURE TO UPDATE NAMES OF THE SAVED OUTPUT FILES ########
# main.data = read.csv("../Data/Step1_PrepareFeatures_Baseline.csv")

source("../R_Helper/Step3_helper.R")
main.data.20 = variable_prep(main.data)

pander(names(main.data.20))
# head(main.data.20)
```

* Relies heavily on "Step3_helper.R." This code does the following:

  * remove levels that are fewer than 20 turns
  * Convert # of turns to a running proportion score 
  * Get z-scores for relevant variables
  * create ordered factor for block (interested in linear trend)
  * create random factor for within subject variance (making sure each unique subject is signified given their school and team)

### Step 2: Check the distributions of the main outcome linguistic variables

```{r, echo=FALSE}
g1 = main.data.20 %>% select(lexical, syntax, semantic) %>% pivot_longer(lexical:semantic, "align") %>% mutate(align=factor(align)) %>% mutate(align=fct_relevel(align, c("lexical", "syntax", "semantic")))
align.labels = c("Lexical", "Syntax", "Semantic")
names(align.labels) = c("lexical", "syntax", "semantic")
ggplot(g1, aes(x=value))+
  geom_histogram(binwidth=0.05)+
  facet_wrap(~align, labeller = labeller(align = align.labels), scales = "free")+
  # labs(title="Distribution of Alignment Scores",x="Align Value (Cosine)", y = "Frequency") +
  labs(y = "Frequency") +
  theme_bw() + 
  theme(plot.title=element_text(size=14,face="bold"), # theme_light()
  #       axis.text.y=element_text(size=12),
        axis.ticks.y=element_blank(),
  #       axis.text.x=element_text(size=12, face="bold"),
  #       axis.title=element_text(size=12,face="bold"),
  #       # plot.caption = element_text(hjust = 0),
  #       # strip.text.y = element_text(hjust=0,vjust = 1,angle=180,face="bold"),
    axis.line = element_line(color = "#3D4852"),
    axis.ticks = element_line(color = "#3D4852"),
    panel.grid.major.y = element_line(color = "#DAE1E7"),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.y = element_blank()
  )
# ggsave("Figs/histogram.pdf")
# ggsave("Figs/histogram_baseline.pdf")
```

```{r, echo=FALSE}
# ggsave(filename = "Figs/histogram.png",width = 6, height = 3, dpi = 300, units = "in", device='png')
```

```{r, echo=FALSE}
# What is just the overall amount of alignment for each linguistic type?
# summtab = main.data.20 %>% 
#   select(Block, lexical, syntax, semantic) %>%
#   group_by(Block) %>%
#   summarise(across(everything(), .f = list(mean = mean, sd = sd), na.rm = TRUE))
# pander(summtab, digits=3)
```

```{r}
g1 = main.data.20 %>% select(lexical, syntax, semantic)
head(g1)
```

<!-- Logistic regression coefficients come out on the logit scale -->
<!-- Gamma coefficients are on a log-scale  -->
<!-- Both can be exponentiated to produce odds-ratios.  -->

<!-- For a gamma model with a log link, you would be modelling the log expected value of a (strictly positive) continuous response variable but the interpretation of the regression coefficient for a predictor would be similar to what I described above. -->

### Step 3: Build and test models: Lexical

<!-- Our solution for addressing the excess of zeros in the data is to separately model these from the non-zeros in a binomial-Gamma hurdle model. -->

```{r, results='hide'}
# library(glmmTMB) ## no longer using
library(lme4) ## for glmer 
library(strengejacke) ## for tab_model and plot_model
library(parameters)
library(lmtest) ## for lrtest
library(bbmle) ## for AICtab
```

Setup for separating prevalence from magnitude

```{r}
GHM.data.20 = main.data.20 %>% mutate(non_zero_lex = ifelse(lexical > 0, 1, 0))
```
<!-- ## we want to predict nonzero values, therefore must set reference to ZERO (glmer does this by default) -->
<!-- # GHM.data.20$non_zero_lex = relevel(factor(GHM.data.20$non_zero_lex), ref="ZERO") -->

####################################################################################################
####################################################################################################

#### Binomial logistic model 

##### Model selection tests: interceptVSslope

To start, lets examine the binomial logistic model by first comparing two plausible random effect structures (while avoiding convergence issues).

Run the models: 

```{r}
z1.glmer1 <- glmer(non_zero_lex ~ ordering_prop + Block.l + aligner_role + relative_start_timeZ + level_durationZ + utterlen_alignerZ + Concept + (1|subject_id), binomial(link = "logit"), data = GHM.data.20)
```

```{r}
z1.glmer2 <- glmer(non_zero_lex ~ ordering_prop + Block.l + aligner_role + relative_start_timeZ + level_durationZ + utterlen_alignerZ + Concept + (aligner_role|subject_id), binomial(link = "logit"), data = GHM.data.20)
```

Compare the two models based on comparing log-likelihood/AIC with various tests. 

```{r}
# AICtab
AICtab(z1.glmer1,z1.glmer2,logLik=TRUE)
```

```{r}
# lrtest
lrtest(z1.glmer1,z1.glmer2)
```

##### Model selection tests: randomVSnone

Lets examine the summary structure of the preferred model, now focusing on the appropriateness of the random effects of the model. It does seem that the role|subject variance is very small. Is this variance relevant? Next step is to compare this effect with a model where we pool/ignore the role|subject effect. 

```{r}
summary(z1.glmer2)
```

The best that can be said is that the random and pooled have non identical log-likelihoods, and based on AIC, the random effect structure presents a slightly better model. 

```{r}
z1.glm2 <- glm(non_zero_lex ~ ordering_prop + Block.l + aligner_role + relative_start_timeZ + level_durationZ + utterlen_alignerZ + Concept, binomial(link = "logit"), data = GHM.data.20)
```

```{r}
# AICtab
AICtab(z1.glm2,z1.glmer2,logLik=TRUE)

#Method 2: lrtest : CANNOT RUN THIS TEST BECAUSE TWO DIFFERENT MODEL CLASSES! 
```

<!-- Lastly, for random efffects, let's compute the ICC value -->

```{r, echo=FALSE, results='hide'}
performance::icc(z1.glmer2)
```

##### Model selection tests: fullVsnull

Next, after making some observations about the random effects structure, lets see of the full model is better than the null model. Here we might be on more solid footing in using the LR-test. 

<!-- * Reading in this area that is supportive -->
<!--     * Example of comparing Null and Full with hurdle modeling: https://link.springer.com/article/10.1007/s10940-020-09480-8#ref-CR77 -->
<!--     and https://link.springer.com/article/10.1007/s10940-020-09480-8/tables/4 -->

```{r}
z1.glmer2.null <- glmer(non_zero_lex ~ 1 + (aligner_role|subject_id), binomial(link = "logit"), data = GHM.data.20)
```

```{r}
# AICtab
AICtab(z1.glmer2.null,z1.glmer2,logLik=TRUE)
```

```{r}
# lrtest
lrtest(z1.glmer2.null,z1.glmer2)
```

##### Interpreting

Let's get to interpreting the effects. These results can be reported as "log effect estimates" and exponentiated for "odds ratios." 

<!-- Here are several examples of how these have been reported that seem to make the most sense: -->
<!--   * From Bolker: http://ms.mcmaster.ca/~bolker/misc/GLMM_results_report.pdf -->
<!--   * From extortion victimization paper: https://link.springer.com/article/10.1007/s10940-020-09480-8 -->
<!--   * For generally reporting and intepreting mixed effects models: http://www.singmann.org/download/publications/singmann_kellen-introduction-mixed-models.pdf -->

```{r}
# model_parameters(z1.glmer2, exponentiate = T, digits=3)
```

```{r}
model_parameters(z1.glmer2, exponentiate = F, digits=3)

# Now generating the % change from the non-exponentiated coefficients to facilitate reporting 
(1-(exp(-0.068))) * 100
```


####################################################################################################
####################################################################################################

#### Gamma models

##### Model selection tests: interceptVSslope

```{r}
g1.glmer1 <- glmer(lexical ~ ordering_prop + Block.l + aligner_role + relative_start_timeZ + level_durationZ + utterlen_alignerZ + Concept + (1|subject_id), family = Gamma(link = log), data = subset(GHM.data.20, non_zero_lex == 1), control=glmerControl(optimizer="bobyqa"))
```

```{r}
g1.glmer2 <- glmer(lexical ~ ordering_prop + Block.l + aligner_role + relative_start_timeZ + level_durationZ + utterlen_alignerZ + Concept + (aligner_role|subject_id), family = Gamma(link = log), data = subset(GHM.data.20, non_zero_lex == 1), control=glmerControl(optimizer="bobyqa"))
```

Compare the two models based on comparing log-likelihood/AIC with various tests. 

```{r}
# AICtab
AICtab(g1.glmer1,g1.glmer2,logLik=TRUE)

# lrtest
lrtest(g1.glmer1,g1.glmer2)
```

##### Model selection tests: randomVSnone

Lets examine the summary structure of the preferred model, now focusing on the appropriateness of the random effects of the model. Again, role|subject variance is very small. Is this variance relevant? Next step is to compare this effect with a model where we pool/ignore the role|subject effect. 

```{r}
summary(g1.glmer2)
```

The best that can be said is that the random and pooled have non identical log-likelihoods, and based on AIC, the random effect structure presents a slightly better model. 

```{r}
g1.glm2 <- glm(lexical ~ ordering_prop + Block.l + aligner_role + relative_start_timeZ + level_durationZ + utterlen_alignerZ + Concept, family = Gamma(link = log), data = subset(GHM.data.20, non_zero_lex == 1))
```

```{r}
# AICtab
AICtab(g1.glm2,g1.glmer2,logLik=TRUE)
```

Lastly, for random efffects, let's compute the ICC value

```{r, echo=FALSE, results='hide'}
performance::icc(g1.glmer2)
```

##### Model selection tests: fullVsnull

Next, after making some observations about the random effects structure, lets see of the full model is better than the null model. Here we might be on more solid footing in using the LR-test. 

```{r}
g1.glmer2.null <- glmer(lexical ~ 1 + (aligner_role|subject_id), family = Gamma(link = log), data = subset(GHM.data.20, non_zero_lex == 1), control=glmerControl(optimizer="bobyqa"))
```

```{r}
# AICtab
AICtab(g1.glmer2.null,g1.glmer2,logLik=TRUE)
```

```{r}
# lrtest
lrtest(g1.glmer2.null,g1.glmer2)
```

##### Interpreting

Let's get to interpreting the effects. These results can be reported as "log effect estimates" and exponentiated for "odds ratios." 

```{r}
# model_parameters(g1.glmer2, exponentiate = T, digits=3)
```

```{r}
model_parameters(g1.glmer2, exponentiate = F, digits=3)

# (1-(exp(-0.035))) * 100
```

####################################################################################################
####################################################################################################

### Step 4: Create visuals

Let's start with rate incidence/odds ratio as a forest plot using "plot_model" with modifications

```{r, echo=FALSE, fig.show='hide'}
pz.l = plot_model(z1.glmer2, show.values = TRUE, value.offset = .4,
                  group.terms = c(1,1,2,2,2,2,2),
                  rm.terms = c("Block.l.Q"),  
                  axis.labels = c("ordering_prop"="Utterance order", "Block.l.L"="Block (linear trend)", "aligner_rolecontroller"="Role [controller]", "relative_start_timeZ"="Level start time", "level_durationZ"="Level duration", "utterlen_alignerZ"="Utterance length", "ConceptPoT" = "Concept [PoT]"),
                  # title=c("Lexical Alignment"),
                  dot.size=2,
                  digits = 3) + 
                  labs(title = "Lexical Alignment",
                  subtitle = "Binomial Logistic Mixed Effects")

pz.l = pz.l + ylim(.5, 2.45) + 
  geom_hline(yintercept =1, linetype=3) + 
  theme_bw() +
  theme(
        plot.title=element_text(size=16,face="bold"), # theme_light()
        axis.text.y=element_text(size=10),
        # axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.text.x=element_text(face="bold"),
        axis.title=element_text(size=12,face="bold"),
        strip.text.y = element_text(hjust=0,vjust = 1,angle=180,face="bold"))
pz.l
```

```{r, echo=FALSE, fig.show='hide'}
pg.l = plot_model(g1.glmer2, show.values = TRUE, value.offset = .4,
                  group.terms = c(1,1,2,2,2,2,2),
                  rm.terms = c("Block.l.Q"),  
                  axis.labels = c("ordering_prop"="Utterance order", "Block.l.L"="Block (linear trend)", "aligner_rolecontroller"="Role [controller]", "relative_start_timeZ"="Level start time", "level_durationZ"="Level duration", "utterlen_alignerZ"="Utterance length", "ConceptPoT" = "Concept [PoT]"),
                  title=c(""),
                  dot.size=2,
                  digits = 3) + 
                  labs(title = "",
                  subtitle = "Gamma Mixed Effects")
                  # caption = expression(paste("Significance: ***", italic("p"), "<.001, **", italic("p"), "<.001")))

pg.l = pg.l + ylim(.85, 1.15) + 
  scale_y_continuous(name="Rate Ratios") +
  geom_hline(yintercept =1, linetype=3) + 
  theme_bw() + 
  theme(plot.title=element_text(size=16,face="bold"), # theme_light()
        # axis.text.y=element_text(size=12),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.text.x=element_text(face="bold"),
        axis.title=element_text(size=12,face="bold"),
        # plot.caption = element_text(hjust = 0),
        strip.text.y = element_text(hjust=0,vjust = 1,angle=180,face="bold"))
pg.l 
```

```{r}
library(patchwork)
alignPlotLex = pz.l + pg.l
alignPlotLex
# alignPlotLex + plot_annotation(tag_levels = 'A')
```

```{r, echo=FALSE}
# ggsave(filename = "Figs/lexical_forest.png",width = 8, height = 4.5, dpi = 300, units = "in", device='png')
# ggsave(filename = "Figs/lexical_forest_baseline.png",width = 8, height = 4.5, dpi = 300, units = "in", device='png')

```

```{r, echo=FALSE}
save.image(file = "lexical.RData")
# save.image(file = "lexical_baseline.RData")
```
