==> R_RMD_Analysis/Step3_M2_Trophies.Rmd <==
---
title: "Multi-Level Linguistic Alignment in a Dynamic CPS Tasks: Step 3 Model 2: Trophies"
date: 03/28/22
output: html_document
---

```{r global-options, include=FALSE}
knitr::opts_chunk$set(fig.path='Figs/',
                      warning=FALSE, message=FALSE, error=FALSE)
```

## Notebook contains code for replicating section in manuscript: Analysis 2: Alignment and CPS outcomes

#### Step 1: Import main data file and run script to prep variables for analysis

```{r}
library(tidyverse)
library(pander)

rm(list=ls())
```

```{r}
main.data = read.csv("../Data/Step1_PrepareFeatures.csv")

source("../R_Helper/Step3_helper.R")
main.data.20 = variable_prep(main.data)

pander(names(main.data.20))
# head(main.data.20)
```

* Relies heavily on "Step3_helper.R." This code does the following:

  * remove levels that are fewer than 20 turns
  * Convert # of turns to a running proportion score 
  * Get z-scores for relevant variables
  * create ordered factor for block (interested in linear trend)
  * create random factor for within subject variance (making sure each unique subject is signified given their school and team)

### Step 2: Build and test models: Ordinal regression mixed-effects models

```{r, results='hide'}
library(ordinal)
library(parameters) 
library(strengejacke)
```

Setup for identifying outcomes to be predicted

```{r}
main.data.20$trophyNum <- ifelse(main.data.20$trophy == "none", 1, 
  ifelse(main.data.20$trophy == "silver" | main.data.20$trophy == "trophy", 2, 
  ifelse(main.data.20$trophy == "gold", 3, NA)))
```

```{r, echo=FALSE}
# What is just the overall amount of alignment for each linguistic type?
# summtabtrop = main.data.20 %>% 
#   select(trophy, lexical, syntax, semantic) %>%
#   group_by(trophy) %>%
#   summarise(across(everything(), .f = list(mn = mean, sd = sd), na.rm = TRUE))
# panderOptions('table.split.table', 100)
# pandoc.table(summtabtrop, style = 'simple', digits=3)
```


```{r}
clmm_all <- clmm(as.factor(trophyNum) ~ lexical + syntax + semantic + (center(lexical) + center(syntax) + center(semantic)):center(ordering_prop) + 
              level_durationZ + relative_start_timeZ + revisited + Concept + Block.l + 
              (1|team_id), data = main.data.20, link = "logit")
```

```{r}
model_parameters(clmm_all, digits=3)
performance::icc(clmm_all)
```

```{r, echo=FALSE}
tab_model(clmm_all,
          rm.terms = c("1|2", "2|3", "Block.l.Q"),
          order.terms = c(1,2,3,9,10,11,8,4,5,6,7),
          pred.labels = c("semantic"="Semantic", "syntax"="Syntax", "lexical"="Lexical",
                          "center(semantic):center(ordering_prop)"="Semantic * Time",
                          "center(lexical):center(ordering_prop)"="Lexical * Time",
                          "center(syntax):center(ordering_prop)"="Syntactic * Time",                         
                          "Block.l.L"="Block (linear trend)",
                          "level_durationZ"="Level Duration",
                          "relative_start_timeZ"="Level Start Time",
                          "ConceptPoT" = "Concept [PoT]", 
                          "revisitedreattempt" = "Revisit [yes]"), 
          digits = 3,
          digits.re = 3,
          dv.labels = c("Trophy"),
          show.re.var = FALSE,
          show.r2 = FALSE,
          show.icc = FALSE,
          show.ngroups = FALSE,
          show.obs = FALSE,
          use.viewer = TRUE)
```

##### Lexical

<!-- https://cran.r-project.org/web/packages/ordinal/vignettes/clmm2_tutorial.pdf -->

```{r}
clmm_lex <- clmm(as.factor(trophyNum) ~ lexical + center(lexical):center(ordering_prop) + 
              level_durationZ + relative_start_timeZ + revisited + Concept + Block.l + 
              (1|team_id), data = main.data.20, link = "logit", threshold="flexible")
```

```{r}
model_parameters(clmm_lex, digits=3)
performance::icc(clmm_lex)
```

Interpretation: output is odds ratios, values above 1 indicate a greater likelihood of receiving a trophy, values below 1 indicate less likelihood of receiving a trophy 

```{r, echo=FALSE}
tab_model(clmm_lex,
          
          rm.terms = c("1|2", "2|3", "Block.l.Q"),
          # order.terms = c(3,10,8,4,5,6,7),
          order.terms = c(1,7,6,2,3,4,5),
          pred.labels = c("lexical"="Lexical", "center(lexical):center(ordering_prop)"="Lexical * Order", "Block.l.L"="Block (linear trend)", "level_durationZ"="Level duration", "relative_start_timeZ"="Level start time", "revisitedreattempt" = "Revisit [yes]",  "ConceptPoT" = "Concept [PoT]"),          
          title = "Lexical Alignment",
          digits = 3,
          digits.re = 3,
          dv.labels = c("Trophy"),
          show.re.var = FALSE,
          show.r2 = FALSE,
          show.icc = FALSE,
          show.ngroups = FALSE,
          show.obs = FALSE,
          use.viewer = TRUE)
```

##### Syntactic

```{r}
clmm_syn <- clmm(as.factor(trophyNum) ~ syntax + center(syntax):center(ordering_prop) +  
              level_durationZ + relative_start_timeZ + revisited + Concept + Block.l + 
              (1|team_id), data = main.data.20, link = "logit")
```

```{r}
model_parameters(clmm_syn, digits=3)
# performance::icc(clmm_syn)
```

```{r, echo=FALSE}
tab_model(clmm_syn,
          
          rm.terms = c("1|2", "2|3", "Block.l.Q"),
          # order.terms = c(3,10,8,4,5,6,7),
          order.terms = c(1,7,6,2,3,4,5),
          pred.labels = c("syntax"="Syntax", "center(syntax):center(ordering_prop)"="Syntax * Order", "Block.l.L"="Block (linear trend)", "level_durationZ"="Level duration", "relative_start_timeZ"="Level start time", "revisitedreattempt" = "Revisit [yes]",  "ConceptPoT" = "Concept [PoT]"),
          title = "Syntactic Alignment",
          
          digits = 3,
          digits.re = 3,
          dv.labels = c("Trophy"),
          show.re.var = FALSE,
          show.r2 = FALSE,
          show.icc = FALSE,
          show.ngroups = FALSE,
          show.obs = FALSE,
          use.viewer = TRUE)
```

##### Semantic

```{r}
clmm_sem <- clmm(as.factor(trophyNum) ~ semantic + center(semantic):center(ordering_prop) + 
              level_durationZ + relative_start_timeZ + revisited + Concept + Block.l + 
              (1|team_id), data = main.data.20, link = "logit")
```

```{r}
## still significant without interaction with time
# clmm_sem <- clmm(as.factor(trophyNum) ~ semantic + 
#               level_durationZ + relative_start_timeZ + revisited + Concept + Block.l + 
#               (1|team_id), data = main.data.20, link = "logit")
```

```{r}
model_parameters(clmm_sem, digits=3)
# performance::icc(clmm_sem)
```

```{r, echo=FALSE}
tab_model(clmm_sem,

          rm.terms = c("1|2", "2|3", "Block.l.Q"),
          # order.terms = c(3,10,8,4,5,6,7),
          order.terms = c(1,7,6,2,3,4,5),
          pred.labels = c("semantic"="Semantic", "center(semantic):center(ordering_prop)"="Semantic * Order", "Block.l.L"="Block (linear trend)", "level_durationZ"="Level duration", "relative_start_timeZ"="Level start time", "revisitedreattempt" = "Revisit [yes]",  "ConceptPoT" = "Concept [PoT]"),
          title = "Semantic Alignment",
          
          digits = 3,
          digits.re = 3,
          dv.labels = c("Trophy"),
          show.re.var = FALSE,
          show.r2 = FALSE,
          show.icc = FALSE,
          show.ngroups = FALSE,
          show.obs = FALSE,
          use.viewer = TRUE)
```

```{r}
save.image(file = "trophies.RData")
```
